Make realtime extraction layout idiomatic
