+++ codex-rs/realtime/src/lib.rs
@@ -1,25 +1,27 @@
 mod config;
+mod handoff;
 mod prompt;
 mod realtime_conversation;
+mod session_config;
 mod startup_context;
 
 pub use config::RealtimeFeaturesConfig;
+pub use handoff::REALTIME_BACKEND_TEXT_PREFIX;
+pub use handoff::REALTIME_USER_TEXT_PREFIX;
+pub use handoff::prefix_realtime_v2_text;
+pub use handoff::realtime_delegation_from_handoff;
+pub use handoff::realtime_text_from_handoff_request;
+pub use handoff::wrap_realtime_delegation_input;
 pub use prompt::prepare_realtime_backend_prompt;
-pub use realtime_conversation::DEFAULT_REALTIME_MODEL;
-pub use realtime_conversation::REALTIME_BACKEND_TEXT_PREFIX;
-pub use realtime_conversation::REALTIME_USER_TEXT_PREFIX;
 pub use realtime_conversation::RealtimeConversationManager;
 pub use realtime_conversation::RealtimeStart;
 pub use realtime_conversation::RealtimeStartOutput;
-pub use realtime_conversation::build_realtime_session_config;
-pub use realtime_conversation::default_realtime_voice;
-pub use realtime_conversation::prefix_realtime_v2_text;
-pub use realtime_conversation::realtime_api_key;
-pub use realtime_conversation::realtime_delegation_from_handoff;
-pub use realtime_conversation::realtime_request_headers;
-pub use realtime_conversation::realtime_text_from_handoff_request;
-pub use realtime_conversation::validate_realtime_voice;
-pub use realtime_conversation::wrap_realtime_delegation_input;
+pub use session_config::DEFAULT_REALTIME_MODEL;
+pub use session_config::build_realtime_session_config;
+pub use session_config::default_realtime_voice;
+pub use session_config::realtime_api_key;
+pub use session_config::realtime_request_headers;
+pub use session_config::validate_realtime_voice;
 pub use startup_context::CURRENT_THREAD_SECTION_TOKEN_BUDGET;
 pub use startup_context::NOTES_SECTION_TOKEN_BUDGET;
 pub use startup_context::REALTIME_TURN_TOKEN_BUDGET;
