+++ codex-rs/realtime/src/realtime_conversation_tests.rs
@@ -1,8 +1,5 @@
 use super::RealtimeHandoffState;
 use super::RealtimeSessionKind;
-use super::realtime_delegation_from_handoff;
-use super::realtime_text_from_handoff_request;
-use super::wrap_realtime_delegation_input;
 use async_channel::bounded;
 use codex_protocol::protocol::RealtimeHandoffRequested;
 use codex_protocol::protocol::RealtimeTranscriptEntry;
@@ -26,7 +23,7 @@ fn prefers_handoff_input_transcript_over_active_transcript() {
         ],
     };
     assert_eq!(
-        realtime_text_from_handoff_request(&handoff),
+        crate::realtime_text_from_handoff_request(&handoff),
         Some("ignored".to_string())
     );
 }
@@ -43,7 +40,7 @@ fn extracts_text_from_handoff_request_active_transcript_if_input_missing() {
         }],
     };
     assert_eq!(
-        realtime_text_from_handoff_request(&handoff),
+        crate::realtime_text_from_handoff_request(&handoff),
         Some("user: hello".to_string())
     );
 }
@@ -66,7 +63,7 @@ fn wraps_handoff_with_transcript_delta() {
         ],
     };
     assert_eq!(
-        realtime_delegation_from_handoff(&handoff),
+        crate::realtime_delegation_from_handoff(&handoff),
         Some(
             "<realtime_delegation>\n  <input>delegate this</input>\n  <transcript_delta>user: hello\nassistant: hi there</transcript_delta>\n</realtime_delegation>"
                 .to_string()
@@ -83,7 +80,7 @@ fn extracts_text_from_handoff_request_input_transcript_if_messages_missing() {
         active_transcript: vec![],
     };
     assert_eq!(
-        realtime_text_from_handoff_request(&handoff),
+        crate::realtime_text_from_handoff_request(&handoff),
         Some("ignored".to_string())
     );
 }
@@ -96,13 +93,13 @@ fn ignores_empty_handoff_request_input_transcript() {
         input_transcript: String::new(),
         active_transcript: vec![],
     };
-    assert_eq!(realtime_text_from_handoff_request(&handoff), None);
+    assert_eq!(crate::realtime_text_from_handoff_request(&handoff), None);
 }
 
 #[test]
 fn wraps_realtime_delegation_input() {
     assert_eq!(
-        wrap_realtime_delegation_input("hello", /*transcript_delta*/ None),
+        crate::wrap_realtime_delegation_input("hello", /*transcript_delta*/ None),
         "<realtime_delegation>\n  <input>hello</input>\n</realtime_delegation>"
     );
 }
@@ -110,7 +107,7 @@ fn wraps_realtime_delegation_input() {
 #[test]
 fn wraps_realtime_delegation_input_with_xml_escaping() {
     assert_eq!(
-        wrap_realtime_delegation_input("use a < b && c > d", Some("saw <that>")),
+        crate::wrap_realtime_delegation_input("use a < b && c > d", Some("saw <that>")),
         "<realtime_delegation>\n  <input>use a &lt; b &amp;&amp; c &gt; d</input>\n  <transcript_delta>saw &lt;that&gt;</transcript_delta>\n</realtime_delegation>"
     );
 }
@@ -118,7 +115,7 @@ fn wraps_realtime_delegation_input_with_xml_escaping() {
 #[test]
 fn wraps_realtime_delegation_input_with_xml_escaping_without_transcript() {
     assert_eq!(
-        wrap_realtime_delegation_input("use a < b && c > d", /*transcript_delta*/ None),
+        crate::wrap_realtime_delegation_input("use a < b && c > d", /*transcript_delta*/ None),
         "<realtime_delegation>\n  <input>use a &lt; b &amp;&amp; c &gt; d</input>\n</realtime_delegation>"
     );
 }
